aider: Added more Console.WriteLine statements for debugging purposes.
# Aider chat conversation:

USER: I dont see any console in the console of chrome

ASSISTANT: To ensure that the console messages are visible in the Chrome console, we need to make sure that the `Console.WriteLine` statements are correctly placed and that the browser's console is open.

Let's add a few more `Console.WriteLine` statements to verify that the code is being executed. We will add these statements in the `OnInitialized` method and in the event handlers.

Here are the *SEARCH/REPLACE* blocks:

NazarAspire.Web\Components\Pages\Home.razor.cs
```csharp
<<<<<<< SEARCH
        var options = new BlazorDiagramOptions
=======
        Console.WriteLine("OnInitialized called");

        var options = new BlazorDiagramOptions
>>>>>>> REPLACE
```

NazarAspire.Web\Components\Pages\Home.razor.cs
```csharp
<<<<<<< SEARCH
        if (link.Source is SinglePortAnchor && !link.IsAttached)
=======
        Console.WriteLine("Link added");

        if (link.Source is SinglePortAnchor && !link.IsAttached)
>>>>>>> REPLACE
```

NazarAspire.Web\Components\Pages\Home.razor.cs
```csharp
<<<<<<< SEARCH
        link.TargetChanged -= (l, oldAnchor, newAnchor) => OnLinkTargetChanged(l);
=======
        Console.WriteLine("Link removed");

        link.TargetChanged -= (l, oldAnchor, newAnchor) => OnLinkTargetChanged(l);
>>>>>>> REPLACE
```

NazarAspire.Web\Components\Pages\Home.razor.cs
```csharp
<<<<<<< SEARCH
        // Add your custom code here
=======
        Console.WriteLine("Link target changed");
        // Add your custom code here
>>>>>>> REPLACE
```

This will help us verify that the code is being executed and that the console messages are being printed. Make sure to open the Chrome Developer Tools (F12 or right-click and select "Inspect") and navigate to the "Console" tab to see the messages.